requirements-dev.txt
# For tools/render_solution_images.py
pillow>=9.0
